Remove possibility to pause dialogs
Before, dialog models where not disposed properly.
Adding support for disposing of dialog models would be too much
effort to implement properly. Further, it would make the code
less simple to understand. Given that this will be a rarely used
feature, the feature is removed instead of being properly implemented.